Fixed OpenGLView on Linux
