+++ Xamarin.Forms.ControlGallery.GTK/Program.cs
@@ -20,8 +20,8 @@ namespace Xamarin.Forms.ControlGallery.GTK
             Forms.Init();
             FormsMaps.Init(string.Empty);
 
-            //var app = new Controls.App();
-            var app = new OpenGLApp();
+            var app = new Controls.App();
+            //var app = new OpenGLApp();
             var window = new FormsWindow();
             window.LoadApplication(app);
             window.SetApplicationTitle("Xamarin.Forms GTK# Backend");
