--- Xamarin.Forms.Platform.GTK/Libs/GLWidget/README.txt
@@ -1,5 +1,0 @@
-﻿////////////////////////////////////////////////////////////////////////////////
-// Gtk GLWidget Sharp - Gtk OpenGL Widget for CSharp using OpenTK
-////////////////////////////////////////////////////////////////////////////////
-
-https://sourceforge.net/projects/glwidget/
